Add boutique and collections pages with shared catalogue
