+++ store-frontend/app/boutique/page.tsx
@@ -1,0 +1,5 @@
+import { BoutiqueCatalogue } from '@/components/BoutiqueCatalogue';
+
+export default function BoutiquePage() {
+  return <BoutiqueCatalogue />;
+}
